oracle constants moved to a separate constants module
