+++ kimchi/src/tests/poseidon.rs
@@ -19,7 +19,7 @@ use mina_curves::pasta::{
     vesta::{Affine, VestaParameters},
 };
 use oracle::{
-    poseidon::{PlonkSpongeConstantsKimchi, SpongeConstants},
+    constants::{PlonkSpongeConstantsKimchi, SpongeConstants},
     sponge::{DefaultFqSponge, DefaultFrSponge},
 };
 use rand::{rngs::StdRng, SeedableRng};
